ai/requirements.txt
# Core LangChain
langchain==0.1.0
langchain-core==0.1.10

# For OpenAI (if using ChatGPT)
langchain-openai==0.0.5
openai==1.10.0

# For Google Gemini (if using Gemini instead)
langchain-google-genai==0.0.6
google-generativeai==0.3.2

# Data validation
pydantic==2.5.3

# Environment variables
python-dotenv==1.0.0

# Optional but useful
requests==2.31.0